requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.13.3

# Redis
redis[hiredis]==5.2.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Config
pydantic-settings==2.5.2
python-multipart==0.0.12

# HTTP client (for financialdatasets.ai)
httpx==0.27.2

# Free market data
yfinance==0.2.44
pandas==2.2.3

# Monitoring
python-json-logger==2.0.7
